+++ BET_elimination.cuh
@@ -9,4 +9,4 @@ __device__ int split_poly(int * poly, int length_poly, int * triangles, int * ad
 
 
 __global__ void polygon_reparation(int* cu_mesh, int* cu_mesh_aux, int num_poly, int *cu_ind_poly, int *cu_ind_poly_aux, int *cu_triangles, int tnumber, int *cu_adj, double *cu_r, int *cu_i_mesh, int* cu_i_ind_poly, int *is_there_bet);
-__global__ void polygon_reparation2(int* cu_mesh, int* cu_mesh_aux, int num_poly, int *cu_ind_poly, int *cu_ind_poly_aux, int *cu_triangles, int tnumber, int *cu_adj, double *cu_r, int *cu_i_mesh, int* cu_i_ind_poly, int* cu_trivertex, int *is_there_bet);
+__global__ void polygon_reparation2(int* cu_mesh, int* cu_mesh_aux, int num_poly, int *cu_ind_poly, int *cu_ind_poly_aux, int *cu_triangles, int tnumber, int *cu_adj, double *cu_r, unsigned long long int *cu_i_mesh, unsigned long long int* cu_i_ind_poly, int* cu_trivertex, int *is_there_bet);
